more changes to zip-code as a result of restructuring
