+++ zip-search/src/App.css
@@ -1,3 +1,6 @@
-.body-bg-color{
+.App-header {
   background-color: #222;
+  padding: 20px;
+  color: white;
+  text-align: center;
 }
