+++ zip-search/src/App.js
@@ -1,27 +1,119 @@
-import React, { useState } from 'react';
-import Zip from './zip-search';
-import City from './city-search';
+import React, { Component } from 'react';
 import './App.css';
 
-const App = () => {
-  const [view, setView] = useState(0);
+function City({ location, state, long, lat, population, wages }) {
+    return (
+        <div className="card">
+            <div className="card-header font-weight-bold">
+                {location}
+            </div>
+            <div className="card-body">
+                <ul>
+                    <li className="card-text">State: {state}</li>
+                    <li className="card-text">Location: ({lat}, {long})</li>
+                    <li className="card-text">Population (estimated): {population ? population : 'unknown'}</li>
+                    <li className="card-text">Total Wages: {wages ? wages : 'unknown'}</li>
+                </ul>
+            </div>
+        </div>
+    );
+}
 
-  const handleClick = () => {
-    if (view === 0) {
-      setView(1);
-    } else {
-      setView(0);
+function ZipSearchField({ onChange }) {
+    return (
+        <div className="my-4 d-flex justify-content-center">
+            <div className="input-group">
+                <div className="input-group-prepend">
+                    <span className="input-group-text">Zip Code</span>
+                </div>
+                <input className="form-control" type='text' onChange={onChange} />
+            </div>
+        </div>
+    )
+}
+
+
+class App extends Component {
+    constructor(props) {
+        super(props);
+        this.state = {
+            zipCode: '',
+            cityData: []
+        }
     }
-  }
 
-  return (
-    <div>
-      <div className="d-flex py-4 justify-content-center body-bg-color">
-        <button className="btn btn-light" type="button" onClick={handleClick}>Click here to do a {view === 0 ? 'City-Search' : 'Zip-Search'}</button>
-      </div>
-      {view === 0 ? <Zip /> : <City />}
-    </div>
-  );
+    async zipChanged(e) {
+        let zip = e.target.value;
+        let digitReg = /^[0-9]+$/; //using regex with the suggestion of a classmate
+
+        //match method returns array if match, null if not. with the regex here, it will return null if it is not a digit.
+        if (zip.length === 5 && zip.match(digitReg) !== null) {
+            await fetch(`http://ctp-zip-api.herokuapp.com/zip/${zip}`)
+                .then(response => {
+                    if (!response.ok) {
+                        throw new Error('Invalid Zip Code');
+                    }
+                    return response.json();
+                })
+                .then(success => {
+                    this.setState({
+                        cityData: success
+                    })
+                })
+                .catch(err => {
+                    this.setState({ cityData: [] });
+                    console.log(err);
+                });
+
+            this.setState({
+                zipCode: zip
+            })
+        } else {
+            this.setState({ cityData: [] });
+        }
+    }
+
+    //make the city display dynamically
+    render() {
+        return (
+            <div className="App">
+                <div className="App-header">
+                    <h2>Zip Code Search</h2>
+                </div>
+                <div className="container">
+                    <div className="row d-flex justify-content-center">
+                        <div className="col-lg-6 mx-auto">
+                            <ZipSearchField onChange={e => this.zipChanged(e)} />
+                        </div>
+                    </div>
+                    {this.state.cityData.length > 0 ?
+                        this.state.cityData.map((item) => {
+                            return (
+                                <div className="row mb-3 d-flex justify-content-center" key={item.RecordNumber}>
+                                    <div className="col-lg-6 mx-auto">
+                                        <City
+                                            location={item.LocationText}
+                                            long={item.Long}
+                                            lat={item.Lat}
+                                            state={item.State}
+                                            population={item.EstimatedPopulation}
+                                            wages={item.TotalWages} />
+                                    </div>
+                                </div>
+                            )
+                        }
+                        )
+                        :
+                        <div className="row d-flex justify-content-center">
+                            <div className="col-lg-6">
+                                <h4 className="text-center">No results</h4>
+                            </div>
+                        </div>
+                    }
+                </div>
+            </div>
+        );
+    }
 }
 
 export default App;
