--- zip-search/src/city-search.js
@@ -1,14 +1,0 @@
-import React from 'react';
-import './zip-search.css';
-
-const CitySearch = () => {
-  return(
-    <div>
-      <div className="App-header">
-        <h2>City Search</h2>
-      </div>
-    </div>
-  );
-}
-
-export default CitySearch;
