--- zip-search/src/zip-search.css
@@ -1,6 +1,0 @@
-.App-header {
-  background-color: #222;
-  padding: 20px;
-  color: white;
-  text-align: center;
-}
